moved MsgBlockParser
